progression...

+++ src/contracts/CuraAnnonae.sol
@@ -64,7 +64,7 @@ interface ERC20 {
   function balanceOf(address who) external view returns (uint256);
   function allowance(address owner, address spender) external  view returns (uint256);
   function transfer(address to, uint value) external  returns (bool success);
-  function transferFrom(address from, address to, uint256 value) external returns (bool success);
+  function transferFrom(address from, address to, uint value) external returns (bool success);
   function approve(address spender, uint value) external returns (bool success);
 }
